add settings + SelectSeting
