+++ Assets/Scripts/ScriptableObjects/Settings.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+[CreateAssetMenu(menuName = "ScriptableObjects/Settings")]
+public class Settings : ScriptableObject
+{
+    public Language curentLanguage;
+}
