+++ Assets/Scripts/ScriptableObjects/Settings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4ca86f9a622217f40a46afff667d3fcf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
