+++ Assets/Scripts/SelectSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2b287506c4d1200488267a60a10740fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
