+++ Assets/Seting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39dfadb92f344d94888bdc8dedae173f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
