+++ Assets/Seting/Settings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c59aff7649dd3b48851a8287bd7ff76
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
